refactor(kernel): isolate capsule visibility
